fix: removed the mfa otp grant from token
I was ment to remove this when I removed the original password grant
login flow. The MFA OTP should only be done with the authorization code
grant flow as it is user only and not for client credentials. Leaving it
in caused a security bypass that I wasnt a fan of either.